ENH: Support display of Slice intersection in subset of the Slicer viewers
This is useful when all slices need to be visible but only some of slice
intersection should be displayed.

This commit updates the ModelSlice displayable manager to check if a
given slice intersection model was explicitly associated with a view.

It also updates the slice logic so that it resets the Slice display node
View Node IDs only considering the 3D view node IDs set using method like
vtkMRMLSliceNode::AddThreeDViewID

For example, the following snippet allows to keep all Slice visible in
3D while showing the slice intersection in only one slice viewer:

```
for name in ["Yellow", "Red", "Green"]:
  model_display_node = getNode("%s Display" % name)
  model_display_node.AddViewNodeID("vtkMRMLSliceNodeRed")

  slice_node = getNode("vtkMRMLSliceNode%s" % name)
  slice_node.AddThreeDViewID("vtkMRMLViewNode1")
```